Move Utility into a package and fix tests
