+++ hedera-mirror-monitor/src/main/java/com/hedera/mirror/monitor/publish/transaction/account/CryptoTransferTransactionSupplier.java
@@ -38,8 +38,8 @@ import com.hedera.hashgraph.sdk.Hbar;
 import com.hedera.hashgraph.sdk.NftId;
 import com.hedera.hashgraph.sdk.TokenId;
 import com.hedera.hashgraph.sdk.TransferTransaction;
-import com.hedera.mirror.monitor.Utility;
 import com.hedera.mirror.monitor.publish.transaction.TransactionSupplier;
+import com.hedera.mirror.monitor.util.Utility;
 
 @Data
 public class CryptoTransferTransactionSupplier implements TransactionSupplier<TransferTransaction> {
