+++ hedera-mirror-monitor/src/main/java/com/hedera/mirror/monitor/publish/transaction/schedule/ScheduleSignTransactionSupplier.java
@@ -24,11 +24,11 @@ import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
 import lombok.Data;
 
-import com.hedera.mirror.monitor.Utility;
 import com.hedera.hashgraph.sdk.Hbar;
 import com.hedera.hashgraph.sdk.ScheduleId;
 import com.hedera.hashgraph.sdk.ScheduleSignTransaction;
 import com.hedera.mirror.monitor.publish.transaction.TransactionSupplier;
+import com.hedera.mirror.monitor.util.Utility;
 
 @Data
 public class ScheduleSignTransactionSupplier implements TransactionSupplier<ScheduleSignTransaction> {
