+++ hedera-mirror-monitor/src/main/java/com/hedera/mirror/monitor/publish/transaction/token/TokenUnfreezeTransactionSupplier.java
@@ -22,9 +22,6 @@ package com.hedera.mirror.monitor.publish.transaction.token;
 
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
-
-import com.hedera.mirror.monitor.Utility;
-
 import lombok.Data;
 
 import com.hedera.hashgraph.sdk.AccountId;
@@ -32,6 +29,7 @@ import com.hedera.hashgraph.sdk.Hbar;
 import com.hedera.hashgraph.sdk.TokenId;
 import com.hedera.hashgraph.sdk.TokenUnfreezeTransaction;
 import com.hedera.mirror.monitor.publish.transaction.TransactionSupplier;
+import com.hedera.mirror.monitor.util.Utility;
 
 @Data
 public class TokenUnfreezeTransactionSupplier implements TransactionSupplier<TokenUnfreezeTransaction> {
