+++ hedera-mirror-monitor/src/main/java/com/hedera/mirror/monitor/subscribe/grpc/GrpcClientSDK.java
@@ -25,9 +25,6 @@ import java.time.Instant;
 import java.util.List;
 import java.util.Map;
 import javax.inject.Named;
-
-import com.hedera.mirror.monitor.Utility;
-
 import lombok.extern.log4j.Log4j2;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
@@ -41,6 +38,7 @@ import com.hedera.hashgraph.sdk.TopicMessageQuery;
 import com.hedera.mirror.monitor.MonitorProperties;
 import com.hedera.mirror.monitor.subscribe.SubscribeProperties;
 import com.hedera.mirror.monitor.subscribe.SubscribeResponse;
+import com.hedera.mirror.monitor.util.Utility;
 
 @Log4j2
 @Named
